--- ConsoleApplication1.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.271
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ConsoleApplication1", "ConsoleApplication1\ConsoleApplication1.vcxproj", "{6E155E2C-3465-41CB-86E7-7DF3A3500E33}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6E155E2C-3465-41CB-86E7-7DF3A3500E33}.Debug|x64.ActiveCfg = Debug|x64
-		{6E155E2C-3465-41CB-86E7-7DF3A3500E33}.Debug|x64.Build.0 = Debug|x64
-		{6E155E2C-3465-41CB-86E7-7DF3A3500E33}.Debug|x86.ActiveCfg = Debug|Win32
-		{6E155E2C-3465-41CB-86E7-7DF3A3500E33}.Debug|x86.Build.0 = Debug|Win32
-		{6E155E2C-3465-41CB-86E7-7DF3A3500E33}.Release|x64.ActiveCfg = Release|x64
-		{6E155E2C-3465-41CB-86E7-7DF3A3500E33}.Release|x64.Build.0 = Release|x64
-		{6E155E2C-3465-41CB-86E7-7DF3A3500E33}.Release|x86.ActiveCfg = Release|Win32
-		{6E155E2C-3465-41CB-86E7-7DF3A3500E33}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A8882021-0811-439A-A1EF-2337960E78B9}
-	EndGlobalSection
-EndGlobal
