+++ OpencvTest.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26403.7
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "OpencvTest", "OpencvTest\OpencvTest.vcxproj", "{D836DC6B-352D-4885-BF9D-4ACA0D271B35}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D836DC6B-352D-4885-BF9D-4ACA0D271B35}.Debug|x64.ActiveCfg = Debug|x64
+		{D836DC6B-352D-4885-BF9D-4ACA0D271B35}.Debug|x64.Build.0 = Debug|x64
+		{D836DC6B-352D-4885-BF9D-4ACA0D271B35}.Debug|x86.ActiveCfg = Debug|Win32
+		{D836DC6B-352D-4885-BF9D-4ACA0D271B35}.Debug|x86.Build.0 = Debug|Win32
+		{D836DC6B-352D-4885-BF9D-4ACA0D271B35}.Release|x64.ActiveCfg = Release|x64
+		{D836DC6B-352D-4885-BF9D-4ACA0D271B35}.Release|x64.Build.0 = Release|x64
+		{D836DC6B-352D-4885-BF9D-4ACA0D271B35}.Release|x86.ActiveCfg = Release|Win32
+		{D836DC6B-352D-4885-BF9D-4ACA0D271B35}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
